+++ lib/handlers/api.js
@@ -28,7 +28,6 @@ const express = require('express'),
     FormatterHandler = require('./formatting'),
     utils = require('../utils'),
     logger = require('../logger').logger,
-    clientStateClasses = require('../clientstate'),
     clientStateNormalizer = require('../clientstate-normalizer').ClientStateNormalizer;
 
 class ApiHandler {
@@ -67,7 +66,6 @@ class ApiHandler {
         this.handle.get('/formats', formatter.listHandler.bind(formatter));
 
         this.handle.get('/shortlinkinfo/:id', this.shortlinkInfoHandler.bind(this));
-        this.handle.get('/shortlinkinfocode/:id/:sessionid', this.shortlinkInfoCodeHandler.bind(this));
     }
 
     shortlinkInfoHandler(req, res, next) {
@@ -92,38 +90,6 @@ class ApiHandler {
                 next({
                     statusCode: 404,
                     message: `ID "${id}" could not be found`
-                });
-            });
-    }
-
-    shortlinkInfoCodeHandler(req, res, next) {
-        const id = req.params.id;
-        const sessionid = parseInt(req.params.sessionid);
-        this.storageHandler.expandId(id)
-            .then(result => {
-                const config = JSON.parse(result.config);
-
-                let clientstate = false;
-                if (config.content) {
-                    const normalizer = new clientStateNormalizer();
-                    normalizer.fromGoldenLayout(config);
-
-                    clientstate = normalizer.normalized;
-                } else {
-                    clientstate = new clientStateClasses.ClientState(config);
-                }
-
-                const session = clientstate.findSessionById(sessionid);
-                if (!session) throw {msg: `Session ${sessionid} doesn't exist in this shortlink`};
-
-                res.set('Content-Type', 'text/plain');
-                res.end(session.source);
-            })
-            .catch(err => {
-                logger.debug(`Exception thrown when expanding ${id}: `, err);
-                next({
-                    statusCode: 404,
-                    message: `ID "${id}/${sessionid}" could not be found`
                 });
             });
     }
